[runtime/iobuf/pool] rename aligned to buffer
